rename system modules
