+++ zengine/src/platform/mod.rs
@@ -3,7 +3,7 @@ extern crate sdl2;
 use crate::core::Resource;
 use sdl2::VideoSubsystem;
 
-mod system;
-pub use self::system::PlatformSystem;
+mod platform_system;
+pub use self::platform_system::PlatformSystem;
 
 impl Resource for VideoSubsystem {}
